fec: Prevent integer overflow in path metric calculations
(cherry picked from commit 0fdef8493caf39a1dad726cdf4d7a09cdd5476b9)

+++ gr-fec/lib/viterbi/viterbi.cc
@@ -49,7 +49,7 @@
  */
 #define BUTTERFLY(i, sym)                                         \
     {                                                             \
-        int m0, m1;                                               \
+        long m0, m1;                                              \
                                                                   \
         /* ACS for 0 branch */                                    \
         m0 = state[i].metric + mets[sym];          /* 2*i */      \
